Give ResultError subtypes different freshness reduction factors
A failure to measure a relay is not necessarily that relay's fault. Add
(and use) code that allows different ResultError subtypes to reduce a
result's freshness by different amounts. This way, if a bad result was
not likely to be the relay's fault, it won't be penalized very much.

GH: closes #61