Increase verbosity of test output
There's a known-to-fail test in another branch right now. If I didn't
know that and I was relying on Travis to catch the issue, I would not
be happy with the default level of output. (Specifically, the amount
of output in a list of file names is not enough IMHO. It it was a
different type of list, it might be okay)