Find usable destinations; refactor a function to make that easier ...
Since connect_to_destination_over_circuit already determines if a
detination is usable, it made sense to move it and use it in more
places.

Fun fact:
During the process of determining if a destination is usable, a stream
is created to it. Given the same requests Session object, this stream
can be reused (and is by default in requests code) to make additional
requests to the web server. This is why the function is called
connect_to_destination_over_circuit. When first written, determining if
a destination is usbale was the side effect while opening the socket to
the destination was the primary effect.