(Seem to) properly configure setup.py to pull a dev version of stem
I don't really like this as-is because it pins a specific (dev!) version.
In a super ideal world, our target platforms would distribute stem
packages that have the Timeout exception added. But that won't happen
for a while. In an ideal world, stem will have hit 1.7.0 with the Timeout
exception and we could require at least that.

Maybe this isn't so bad though? Anyway, I gotta get these Travis tests to
pass!

GH: ref #162