Make sure requests use the specified proxies
On one hand, for weeks now I've been running sbws and it seems to be
correctly doing downloads over tor. I've monitored existing circs and
streams with a script that queries the control port and what I saw made
sense.

On the other hand, I just tried running sbws on my laptop over a VPN
that blocks resolution of a hostname of mine. I got a proxy error with
a suspicious hostname in the message that led me to believe that
destination usability checks were not actually being done over Tor like
they are supposed to be.

The way Requests tells you to use proxies is by specifying them *not* on
the session object, but as an argument[0]. So do that.

[0]: http://docs.python-requests.org/en/master/user/advanced/#proxies