Logging config overhaul ...
User should configure logging via the [logging] section
in their config.ini.

User can change the log level using --log-level, or in the config
with level, to_file_level, and to_stdout_level.

User can change the log format in the config with format,
to_file_format, and to_stdout_format

An example debugging format is included as a comment in config.default.ini

config.log.default.ini is majorly reorganized. For the most part, if
something doesn't need to be specified in this file, it has been removed.
For example, we don't have to specify any args to handler_to_file. The
beginnings of syslog configuration stuff remain as comments.

As part of configure_logging, force logging to stdout if we don't think
we can log to file.

Actually set some parameters to RotatingFileHandler so it rotates when
files reach 10 MiB and keeps 100 backups.

Each sbws command (like scanner and generate) gets its own log file. I
did this because I didn't want 'sbws scanner' to be running 24/7 and an
'sbws generate' call to log to the same file. That would get confusing.