Stop removing results that are not away
from some other X secs.
In #27338 it was implemented to do not include relays in the
bandwidth files that does not have at least 2 measurements that
are X secs away from each other.
It was implemented removing any results that were not X secs away
from each other, when all the results should be keep when there is
at least one away from the others, as commented in #28061.