globals: remove unused resolve and can_exit_to functions
They were used to resolve the IP of the destination and check
whether an exit policy allows to exit to that IP, but when the
destination is a CDN, the IP locally resolved would be different
to the IP resolved in the exit, and when the IP resolved to
IPv6, it was possible that the scanner didn't have IPv6.
The correct method to check whether an exit policy allows to exit
to an IP, would be to resolve the domain via Tor itself using
RESOLVE and ADDRMAP events with that exit.