config: change default logging levels and formats
And document that when sbws is launched by systemd the stdandard
output will go to journal and syslog.
Set default level to debug when logging to file and a more
verbose format.
Operators don't usually want to see the debug log in the stdout
or syslog, but they can see it in the file.