circuitbuilder: Stop building the circuit 3 times
It is observed that when the circuit can not be built the first
time, consecutive attemps will also fail.
The relay will be measured anyway in the next iteration.
This will also speed up the scanner.
There is no need to remove the 3 attemps in Destination.
_perform_usability_test, since the method will be removed in #28897.

Fixes bug #29295. Bugfix v0.1.0.