chg: scanner: Remove not needed thread loop
The pool can queue all the relays to be measured,
but this inner loop was not adding more worker threads to the pool
unit it finishs the current threads.
Instead, wait for all the worker threads in the pool to finish
or try to get the status after waiting enough.

Closes: #28864, #28865. Bugfix v1.0.0