fix: scripts: Change Github by Gitlab
releases can live now in gitlab.tpo, instead of github.com and
there is no need to check them since Gitlab is FLOSS and gitlab.tpo is
hosted by Tor Project.

Also, stop assuming which is the current branch and remote and do not
push. Instead guide the maintainer to do it.